[combined] Add comments for the important parts
